Rename perfkitbenchmarker_publisher to publisher
